requirements.txt
customtkinter>=5.2.2
pyttsx3>=2.90
SpeechRecognition>=3.10.0
anthropic>=0.34.0
python-dotenv>=1.0.0
pyaudio>=0.2.14
